+++ src/cms_content/cms_content.controller.ts
@@ -1,14 +1,17 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, UploadedFile, UploadedFiles, ParseIntPipe } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, UploadedFile, UploadedFiles, ParseIntPipe, Request, UseGuards } from '@nestjs/common';
 import { CmsContentService } from './cms_content.service';
 import { CreateCmsContentDto } from './dto/create-cms_content.dto';
 import { UpdateCmsContentDto, UpdateJsonContentChildImageDto } from './dto/update-cms_content.dto';
 import { PageType } from './entities/cms_content.entity';
 import { FileFieldsInterceptor, FileInterceptor } from '@nestjs/platform-express';
+import { AuthGuard } from '@nestjs/passport';
+import { ensureCmsDeletePermission } from '../auth/utils/cms-access.util';
 
 @Controller('cms-content')
 export class CmsContentController {
   constructor(private readonly cmsContentService: CmsContentService) {}
 
+  @UseGuards(AuthGuard('jwt'))
   @Post(':page_type')
   @UseInterceptors(FileInterceptor('image'))
   async create(
@@ -30,11 +33,13 @@ export class CmsContentController {
     return this.cmsContentService.findOne(page_type);
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateCmsContentDto: UpdateCmsContentDto) {
     return this.cmsContentService.update(+id, updateCmsContentDto);
-  }
+}
 
+  @UseGuards(AuthGuard('jwt'))
   @Patch('update-with-image/:id')
   @UseInterceptors(FileInterceptor('json_content[mid_image]'))
   async updateWithImage(
@@ -46,6 +51,7 @@ export class CmsContentController {
     return this.cmsContentService.updateWithImage(+id, updateCmsContentDto, imagePath);
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Patch('update-json-child-image/:id')
   @UseInterceptors(FileInterceptor('image'))
   async updateJsonChildImage(
@@ -57,6 +63,7 @@ export class CmsContentController {
     return this.cmsContentService.updateJsonContentChildImage(+id, updateCmsContentDto, imagePath);
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Patch('update-json-homepage-banner/:id')
   @UseInterceptors(FileFieldsInterceptor([
     { name: 'top_icon', maxCount: 1 },
@@ -73,8 +80,10 @@ export class CmsContentController {
     return this.cmsContentService.updateJsonContentHomepageBanner(id, updateCmsContentDto, topIconPath, bannerImagePath);
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @Request() req) {
+    ensureCmsDeletePermission(req.user);
     return this.cmsContentService.remove(+id);
   }
 }
